# docs/swagger.yaml
definitions:
  errorcode.Error:
    type: object
...
  requests.KubePVApplyYamlRequest:
    properties:
      name:
        type: string
      yaml:
        type: string
    type: object
...
  requests.KubePVCreateFromYamlRequest:
    properties:
      yaml:
        type: string
    type: object
  requests.KubePVCreateRequest:
    type: object
  requests.KubePVExpandRequest:
    properties:
      name:
        type: string
      newCapacity:
        description: '例如: 20Gi'
        type: string
    type: object
...
  response.Response:
    type: object
info:
  contact: {}
  description: 基于Gin+Vue开发的K8s管理平台
  termsOfService: https://gitee.com/jay-kim/k8s_operation
  title: K8s管理平台
  version: "1.0"
paths:
...
  /api/v1/k8s/pv/apply-yaml:
    post:
      consumes:
      - application/json
      parameters:
      - description: PV 名称
        in: query
        name: name
        required: true
        type: string
      - description: YAML 内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.KubePVApplyYamlRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 应用 PersistentVolume YAML
      tags:
      - K8s PV 管理
  /api/v1/k8s/pv/create:
    post:
      consumes:
      - application/json
      description: 创建 PV（支持 HostPath / NFS）
      parameters:
      - description: PV 创建参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.KubePVCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 创建 PersistentVolume
      tags:
      - K8s PV 管理
  /api/v1/k8s/pv/create-from-yaml:
    post:
      consumes:
      - application/json
      parameters:
      - description: YAML 内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.KubePVCreateFromYamlRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 从 YAML 创建 PersistentVolume
      tags:
      - K8s PV 管理
  /api/v1/k8s/pv/delete:
    delete:
      parameters:
      - description: PersistentVolume 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 删除 PersistentVolume
      tags:
      - K8s PV 管理
  /api/v1/k8s/pv/detail:
    get:
      parameters:
      - description: PersistentVolume 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 获取 PersistentVolume 详情
      tags:
      - K8s PV 管理
  /api/v1/k8s/pv/expand:
    post:
      consumes:
      - application/json
      description: 扩大 PersistentVolume 容量（只能扩大不能缩小）
      parameters:
      - description: 扩容参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.KubePVExpandRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: PV 扩容
      tags:
      - K8s PV 管理
  /api/v1/k8s/pv/list:
    get:
      description: 支持分页、名称模糊查询（PV 为集群级资源，无 namespace 参数）
      parameters:
      - description: PV 名称(模糊匹配)
        in: query
        maxLength: 100
        name: name
        type: string
      - description: 页码(从1开始)
        in: query
        name: page
        required: true
        type: integer
      - description: 每页数量(默认20)
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 获取 PV 列表
      tags:
      - K8s PV 管理
  /api/v1/k8s/pv/reclaim:
    patch:
      parameters:
      - description: PV 名称
        in: query
        name: name
        required: true
        type: string
      - description: 回收策略 (Delete / Retain)
        in: body
        name: reclaimPolicy
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 修改成功
          schema:
            type: string
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 修改 PersistentVolume 回收策略
      tags:
      - K8s PV 管理
  /api/v1/k8s/pv/yaml:
    get:
      parameters:
      - description: PV 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 获取 PersistentVolume YAML
      tags:
      - K8s PV 管理
...
swagger: "2.0"